REFACT: 파일분리
